+++ Source/NativeApi/NativeApi/Api/Application/KeyEventData.cs
@@ -7,7 +7,7 @@ namespace NativeApi.Api
     public class KeyEventData : NativeEventData
     {
         public Key key;
-        public int timestamp;
+        public long timestamp;
         public bool isRepeat;
     }
 }
